# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  income: "income",
  expense: "expense",
  both: "both",
} as const;

export interface Category {
  id: number;
  name: string;
  icon: string;
  color: string;
  type: CategoryType;
  createdAt: string;
}

export type CreateCategoryBodyType =
  (typeof CreateCategoryBodyType)[keyof typeof CreateCategoryBodyType];

export const CreateCategoryBodyType = {
  income: "income",
  expense: "expense",
  both: "both",
} as const;

export interface CreateCategoryBody {
  name: string;
  icon: string;
  color: string;
  type: CreateCategoryBodyType;
}

export interface IncomeEntry {
  id: number;
  source: string;
  amount: number;
  date: string;
  categoryId: number;
  categoryName: string;
  notes?: string;
  createdAt: string;
}

export interface CreateIncomeBody {
  source: string;
  amount: number;
  date: string;
  categoryId: number;
  notes?: string;
}

export interface ExpenseEntry {
  id: number;
  description: string;
  amount: number;
  date: string;
  categoryId: number;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  notes?: string;
  createdAt: string;
}

export interface CreateExpenseBody {
  description: string;
  amount: number;
  date: string;
  categoryId: number;
  notes?: string;
}

export interface BudgetPlan {
  id: number;
  categoryId: number;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  /** Format: YYYY-MM */
  month: string;
  budgetAmount: number;
  spentAmount: number;
  remainingAmount: number;
  percentUsed: number;
  createdAt: string;
}

export interface CreateBudgetBody {
  categoryId: number;
  /** Format: YYYY-MM */
  month: string;
  budgetAmount: number;
}

export interface SavingsGoal {
  id: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  deadline?: string;
  notes?: string;
  percentComplete: number;
  createdAt: string;
}

export interface CreateSavingsGoalBody {
  name: string;
  targetAmount: number;
  currentAmount: number;
  deadline?: string;
  notes?: string;
}

export interface AddContributionBody {
  amount: number;
  notes?: string;
}

export interface DashboardSummary {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  netSavings: number;
  savingsRate: number;
  budgetUtilization: number;
  topExpenseCategory?: string;
  transactionCount: number;
}

export interface CategoryBreakdown {
  categoryId: number;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  amount: number;
  percentage: number;
  transactionCount: number;
}

export interface MonthlyTrendPoint {
  /** Format: YYYY-MM */
  month: string;
  income: number;
  expenses: number;
  savings: number;
}

export interface BudgetVsActual {
  categoryId: number;
  categoryName: string;
  categoryIcon: string;
  categoryColor: string;
  budgeted: number;
  actual: number;
  variance: number;
  percentUsed: number;
}

export type ListIncomeParams = {
  /**
   * Format: YYYY-MM
   */
  month?: string;
  year?: number;
};

export type ListExpensesParams = {
  /**
   * Format: YYYY-MM
   */
  month?: string;
  categoryId?: number;
};

export type ListBudgetsParams = {
  /**
   * Format: YYYY-MM
   */
  month?: string;
};

export type GetDashboardSummaryParams = {
  /**
   * Format: YYYY-MM. Defaults to current month.
   */
  month?: string;
};

export type GetCategoryBreakdownParams = {
  /**
   * Format: YYYY-MM. Defaults to current month.
   */
  month?: string;
};

export type GetMonthlyTrendParams = {
  months?: number;
};

export type GetBudgetVsActualParams = {
  /**
   * Format: YYYY-MM. Defaults to current month.
   */
  month?: string;
};