# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Budget Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List budget plans
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "Format: YYYY-MM"
      responses:
        "200":
          description: List of budget plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetPlan"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create or update budget for a category/month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetBody"
      responses:
        "201":
          description: Created budget plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetPlan"
  /budgets/{id}:
    put:
      operationId: updateBudget
      tags: [budgets]
      summary: Update budget plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetBody"
      responses:
        "200":
          description: Updated budget plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetPlan"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete budget plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get monthly dashboard summary
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "Format: YYYY-MM. Defaults to current month."
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/category-breakdown:
    get:
      operationId: getCategoryBreakdown
      tags: [dashboard]
      summary: Get expense breakdown by category for a month
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "Format: YYYY-MM. Defaults to current month."
      responses:
        "200":
          description: Category breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
  /dashboard/monthly-trend:
    get:
      operationId: getMonthlyTrend
      tags: [dashboard]
      summary: Get income vs expense trend for the last N months
      parameters:
        - name: months
          in: query
          required: false
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Monthly trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyTrendPoint"
  /dashboard/budget-vs-actual:
    get:
      operationId: getBudgetVsActual
      tags: [dashboard]
      summary: Get budget vs actual spending by category for a month
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "Format: YYYY-MM. Defaults to current month."
      responses:
        "200":
          description: Budget vs actual comparison
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetVsActual"
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - icon
        - color
        - type
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
      required:
        - name
        - icon
        - color
        - type
...
    BudgetPlan:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryIcon:
          type: string
        categoryColor:
          type: string
        month:
          type: string
          description: "Format: YYYY-MM"
        budgetAmount:
          type: number
        spentAmount:
          type: number
        remainingAmount:
          type: number
        percentUsed:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - categoryId
        - categoryName
        - categoryIcon
        - categoryColor
        - month
        - budgetAmount
        - spentAmount
        - remainingAmount
        - percentUsed
        - createdAt
    CreateBudgetBody:
      type: object
      properties:
        categoryId:
          type: integer
        month:
          type: string
          description: "Format: YYYY-MM"
        budgetAmount:
          type: number
      required:
        - categoryId
        - month
        - budgetAmount
...
    DashboardSummary:
      type: object
      properties:
        month:
          type: string
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netSavings:
          type: number
        savingsRate:
          type: number
        budgetUtilization:
          type: number
        topExpenseCategory:
          type: string
        transactionCount:
          type: integer
      required:
        - month
        - totalIncome
        - totalExpenses
        - netSavings
        - savingsRate
        - budgetUtilization
        - transactionCount
    CategoryBreakdown:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryIcon:
          type: string
        categoryColor:
          type: string
        amount:
          type: number
        percentage:
          type: number
        transactionCount:
          type: integer
      required:
        - categoryId
        - categoryName
        - categoryIcon
        - categoryColor
        - amount
        - percentage
        - transactionCount
    MonthlyTrendPoint:
      type: object
      properties:
        month:
          type: string
          description: "Format: YYYY-MM"
        income:
          type: number
        expenses:
          type: number
        savings:
          type: number
      required:
        - month
        - income
        - expenses
        - savings
    BudgetVsActual:
      type: object
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        categoryIcon:
          type: string
        categoryColor:
          type: string
        budgeted:
          type: number
        actual:
          type: number
        variance:
          type: number
        percentUsed:
          type: number
      required:
        - categoryId
        - categoryName
        - categoryIcon
        - categoryColor
        - budgeted
        - actual
        - variance
        - percentUsed