# lib/api-zod/src/generated/types/createBudgetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBudgetBody {
  categoryId: number;
  /** Format: YYYY-MM */
  month: string;
  budgetAmount: number;
}