# lib/api-zod/src/generated/types/monthlyTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Budget Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyTrendPoint {
  /** Format: YYYY-MM */
  month: string;
  income: number;
  expenses: number;
  savings: number;
}